Move helper functions to core_implementations.cc
