+++ shell/platform/common/client_wrapper/include/flutter/basic_message_channel.h
@@ -11,9 +11,28 @@
 
 #include "binary_messenger.h"
 #include "message_codec.h"
-#include "method_channel.h"
 
 namespace flutter {
+
+namespace internal {
+// Internal helper functions used by BasicMessageChannel and MethodChannel.
+
+// Adjusts the number of messages that will get buffered when sending messages
+// to channels that aren't fully set up yet. For example, the engine isn't
+// running yet or the channel's message handler isn't set up on the Dart side
+// yet.
+void ResizeChannel(BinaryMessenger* messenger, std::string name, int new_size);
+
+// Defines whether the channel should show warning messages when discarding
+// messages due to overflow.
+//
+// When |warns| is false, the channel is expected to overflow and warning
+// messages will not be shown.
+void SetChannelWarnsOnOverflow(BinaryMessenger* messenger,
+                               std::string name,
+                               bool warns);
+
+}  // namespace internal
 
 class EncodableValue;
 
@@ -106,7 +125,9 @@ class BasicMessageChannel {
   // to channels that aren't fully set up yet. For example, the engine isn't
   // running yet or the channel's message handler isn't set up on the Dart side
   // yet.
-  void Resize(int new_size) { ResizeChannel(messenger_, name_, new_size); }
+  void Resize(int new_size) {
+    internal::ResizeChannel(messenger_, name_, new_size);
+  }
 
   // Defines whether the channel should show warning messages when discarding
   // messages due to overflow.
@@ -114,7 +135,7 @@ class BasicMessageChannel {
   // When |warns| is false, the channel is expected to overflow and warning
   // messages will not be shown.
   void SetWarnsOnOverflow(bool warns) {
-    SetChannelWarnsOnOverflow(messenger_, name_, warns);
+    internal::SetChannelWarnsOnOverflow(messenger_, name_, warns);
   }
 
  private:
